Get BoVeX compiling with g++ (13.2.0). Still many warnings, some annoying
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5932 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/integer-voronoi.cc
@@ -25,6 +25,7 @@
 
 #include <stdlib.h>
 #include <cstdint>
+#include <cmath>
 
 #include "base/logging.h"
 #include "image.h"
@@ -400,7 +401,7 @@ ImageF IntegerVoronoi::DistanceField(const Image1 &img) {
       int ddx = dx * dx;
       int ddy = dy * dy;
 
-      float d = sqrtf(ddx + ddy);
+      float d = std::sqrt((float)(ddx + ddy));
       /*
       printf("%d,%d nearest %d= %d,%d; dist %.6f\n",
              x, y, p, xx, yy, d);
